Clean up code in preparation for using gcc warnings.

+++ lib/libsys/ds.c
@@ -181,8 +181,8 @@ int ds_retrieve_map(const char *ds_name, char *vaddr, size_t *length,
 			return r;
 
 		/* Do the safemap. */
-		if(*length > m.DS_VAL_LEN)
-			*length = m.DS_VAL_LEN;
+		if(*length > (size_t) m.DS_VAL_LEN)
+			*length = (size_t) m.DS_VAL_LEN;
 		*length = (size_t) CLICK_FLOOR(*length);
 		r = sys_safemap(DS_PROC_NR, m.DS_VAL, 0,
 				(vir_bytes)vaddr, *length, D, 0);
